cafe_location_mvp/requirements.txt
requests
python-dotenv
pydantic
tenacity
